extracted some FXGL deps from MainWindow
